refactor: convert accounts module to table-driven tests and fix clippy warnings
- Create comprehensive table-driven test infrastructure in accounts/common/test_tables.rs
- Share test tables between sync and async implementations for consistency
- Remove duplicate test constants from accounts module (now using common test_utils)
- Fix clippy warnings:
  - Replace redundant continue expressions with empty blocks
  - Nest or-patterns for better readability
- Organize test data with clear separation between test structure and implementation
- All 92 accounts tests pass in both sync and async modes

This improves test maintainability by eliminating code duplication and providing
a single source of truth for test scenarios across sync/async implementations.
