+++ src/accounts/async.rs
@@ -505,7 +505,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_server_version_errors() {
-        use super::common::test_data::tables::VERSION_TEST_CASES;
+        use super::common::test_tables::VERSION_TEST_CASES;
 
         let account = AccountId(TEST_ACCOUNT.to_string());
         let group = AccountGroup("All".to_string());
@@ -538,7 +538,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_managed_accounts_scenarios() {
-        use super::common::test_data::tables::managed_accounts_test_cases;
+        use super::common::test_tables::managed_accounts_test_cases;
 
         for test_case in managed_accounts_test_cases() {
             let (client, message_bus) = if test_case.responses.is_empty() {
@@ -557,7 +557,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_server_time_scenarios() {
-        use super::common::test_data::tables::server_time_test_cases;
+        use super::common::test_tables::server_time_test_cases;
 
         for test_case in server_time_test_cases() {
             let (client, message_bus) = if test_case.responses.is_empty() {
@@ -628,7 +628,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_account_summary_multiple_tags() {
-        use super::common::test_data::tables::account_summary_tag_test_cases;
+        use super::common::test_tables::account_summary_tag_test_cases;
 
         let test_cases = account_summary_tag_test_cases();
 
@@ -694,7 +694,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_pnl_different_model_codes() {
-        use super::common::test_data::tables::pnl_parameter_test_cases;
+        use super::common::test_tables::pnl_parameter_test_cases;
 
         let test_cases = pnl_parameter_test_cases();
         let (client, message_bus) = create_test_client();
@@ -753,7 +753,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_positions_multi_with_various_parameters() {
-        use super::common::test_data::tables::positions_multi_parameter_test_cases;
+        use super::common::test_tables::positions_multi_parameter_test_cases;
 
         let test_cases = positions_multi_parameter_test_cases();
         let (client, message_bus) = create_test_client_with_responses(vec![responses::POSITION_MULTI.into(), responses::POSITION_MULTI_END.into()]);
@@ -821,7 +821,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_subscription_cleanup_on_drop() {
-        use super::common::test_data::tables::{subscription_lifecycle_test_cases, SubscriptionType};
+        use super::common::test_tables::{subscription_lifecycle_test_cases, SubscriptionType};
 
         let test_cases = subscription_lifecycle_test_cases();
         let (client, message_bus) = create_test_client();
@@ -916,7 +916,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_edge_case_contract_ids() {
-        use super::common::test_data::tables::contract_id_test_cases;
+        use super::common::test_tables::contract_id_test_cases;
 
         let test_cases = contract_id_test_cases();
         let (client, message_bus) = create_test_client();
